requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
python-multipart==0.0.12
pdfplumber==0.11.4
python-docx==1.1.2
rapidfuzz==3.10.1
reportlab==4.2.5
pydantic==2.10.3
